Update version and add log interceptor
Updated the package version from v0.0.1 to v0.0.1-dev01 to reflect the development status. Also added a log interceptor to the gRPC utility for better debugging and monitoring. This is noted in the CHANGELOG.md file. Additionally, the new class is exported in the main library file, api_utils.dart.