Add use case results & background execution support
Added support for handling results of use cases and also support for background execution of expensive operations that ideally shouldn't be performed on the main isolate. Created `UseCaseResult` to wrap the result of a `UseCase` execution, updated `UseCase` class and its special types with new features.

Also, updated the library export list, bumped version in pubspec and updated the changelog. Additionally, added a test for the new use case. Equatable package was added to dependencies for object comparison in tests.

This makes our package more robust and flexible in handling different types of operations and improves testing capabilities.