"Added fold method to UseCaseResult"
In this commit, a new fold method has been added to the UseCaseResult class. This allows the wrapping of success and error results for improved flexibility and error handling. As a side effect, responses in the BackgroundUseCase are transferred using UseCaseResult. This could improve efficiency in operations that should not be performed on the main isolate, thus enhancing overall performance.