Update README with package details and usage info
The README has been updated significantly to provide more detailed, useful and structured information about the Flutter package. The new README includes a brief description, usage of the package, and examples of implementation. This change helps prospective users to understand the purpose and functionality of the package, and how to make unary and streaming gRPC calls and HTTP 1.1 requests using Retrofit. The update includes use case templates for implementing Clean Architecture, along with domain, data, and presentation layer examples. Additionally, gRPC and Retrofit helper functions are presented and explained. This is enhancing the documentation and therefore supports user adoption and usage.