Update API call functions to use new FutureEither typedef
This commit includes changes to the `runApiCall`, `runWithGrpcUnaryZonedGuarded`, and `runWithGrpcStreamZonedGuarded` functions in `rest.utils.dart` and `grpc.utils.dart` to use a new typedef `FutureEither` rather than `Future<Either<L, R>>`. This results in a more concise and readable code.

A `typedef.dart` file was created as part of this commit, which includes the new `FutureEither` typedef. The typedefs were then exported through `api_utils.dart`.

Also, the package version was incremented to `0.0.1-dev05` and the `CHANGELOG.md` was updated accordingly.

Overall, these changes are aimed at simplifying the function signatures and making the API more intuitive to use.