Increment package version and swap UseCaseResult fold function
This commit updates the version of the package from 0.0.1-dev05 to 0.0.1-dev06. The change has been noted in the CHANGELOG.md file.

Furthermore, the `fold` function within the UseCaseResult, UseCaseResultSuccess, and UseCaseResultError classes has been modified. The parameters of the `fold` function were swapped to improve usability and meet standard function parameter order of success and error callbacks. This fix is applied to provide a more intuitive API and maintain consistency with common error handling practices.